moved localized composites snippet back
