+++ packages/storybook/stories/snippets/localization/LocalizedComposites.snippet.tsx
@@ -1,0 +1,75 @@
+import { AzureCommunicationTokenCredential } from '@azure/communication-common';
+import {
+  CallComposite,
+  CallAdapter,
+  createAzureCommunicationCallAdapter,
+  ChatComposite,
+  ChatAdapter,
+  createAzureCommunicationChatAdapter,
+  fr_FR
+} from '@azure/communication-react';
+import React, { useEffect, useMemo, useState } from 'react';
+
+function App(): JSX.Element {
+  const endpointUrl = '<Azure Communication Services Resource Endpoint>';
+  const userId = '<Azure Communication Services Identifier>';
+  const displayName = '<Display Name>';
+  const token = '<Azure Communication Services Access Token>';
+
+  //Calling Variables
+  //For Group Id, developers can pass any GUID they can generate
+  const groupId = '<Developer generated GUID>';
+  const [callAdapter, setCallAdapter] = useState<CallAdapter>();
+
+  //Chat Variables
+  const threadId = '<Get thread id from chat service>';
+  const [chatAdapter, setChatAdapter] = useState<ChatAdapter>();
+
+  // We can't even initialize the Chat and Call adapters without a well-formed token.
+  const credential = useMemo(() => {
+    try {
+      return new AzureCommunicationTokenCredential(token);
+    } catch {
+      console.error('Failed to construct token credential');
+      return undefined;
+    }
+  }, [token]);
+
+  useEffect(() => {
+    const createAdapter = async (): Promise<void> => {
+      setChatAdapter(
+        await createAzureCommunicationChatAdapter(
+          endpointUrl,
+          { kind: 'communicationUser', communicationUserId: userId },
+          displayName,
+          new AzureCommunicationTokenCredential(token),
+          threadId
+        )
+      );
+      setCallAdapter(
+        await createAzureCommunicationCallAdapter(
+          { kind: 'communicationUser', communicationUserId: userId },
+          displayName,
+          new AzureCommunicationTokenCredential(token),
+          { groupId }
+        )
+      );
+    };
+    createAdapter();
+  }, []);
+
+  if (!!callAdapter && !!chatAdapter) {
+    return (
+      <>
+        {chatAdapter && <ChatComposite adapter={chatAdapter} locale={fr_FR} />}
+        {callAdapter && <CallComposite adapter={callAdapter} locale={fr_FR} />}
+      </>
+    );
+  }
+  if (credential === undefined) {
+    return <h3>Failed to construct credential. Provided token is malformed.</h3>;
+  }
+  return <h3>Initializing...</h3>;
+}
+
+export default App;
